+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/application/component/ProductEditForm/TabContent/index.ts
@@ -1,9 +1,0 @@
-import DataQualityInsightsTabContent, {CONTAINER_ELEMENT_ID as DATA_QUALITY_INSIGHTS_TAB_CONTENT_CONTAINER_ELEMENT_ID} from './DataQualityInsightsTabContent';
-import AttributesTabContent, {CONTAINER_ELEMENT_ID as ATTRIBUTES_TAB_CONTENT_CONTAINER_ELEMENT_ID} from "./AttributesTabContent";
-
-export {
-  DataQualityInsightsTabContent,
-  DATA_QUALITY_INSIGHTS_TAB_CONTENT_CONTAINER_ELEMENT_ID,
-  AttributesTabContent,
-  ATTRIBUTES_TAB_CONTENT_CONTAINER_ELEMENT_ID
-}