+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/domain/AttributeWithRecommendation.interface.ts
@@ -1,4 +1,0 @@
-export default interface AttributeWithRecommendation {
-  code: string;
-  label: string;
-}