+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/domain/index.ts
@@ -1,26 +1,7 @@
-import Axis, {AxesCollection} from "./Axis.interface";
-import Rates from "./Rates.interface";
-import Rate, {
-  MAX_RATE,
-  RANK_1, RANK_2, RANK_3, RANK_4, RANK_5,
-  RANK_1_COLOR, RANK_2_COLOR, RANK_3_COLOR, RANK_4_COLOR, RANK_5_COLOR, NO_RATE_COLOR
-} from "./Rate.interface";
-import AttributeWithRecommendation from "./AttributeWithRecommendation.interface";
-import Evaluation, {ProductEvaluation, AxisEvaluation, CriterionEvaluationResult} from "./Evaluation.interface";
-import Family from "./Family.interface";
-import Attribute from "./Attribute.interface";
-import Product from "./Product.interface";
-
-export {
-  Axis, AxesCollection,
-  Rate,
-  MAX_RATE,
-  RANK_1, RANK_2, RANK_3, RANK_4, RANK_5,
-  RANK_1_COLOR, RANK_2_COLOR, RANK_3_COLOR, RANK_4_COLOR, RANK_5_COLOR, NO_RATE_COLOR,
-  Rates,
-  AttributeWithRecommendation,
-  Evaluation, ProductEvaluation, AxisEvaluation, CriterionEvaluationResult,
-  Family,
-  Attribute,
+import {
+  CriterionEvaluationResult,
   Product,
-};
+  ProductEvaluation
+} from "@akeneo-pim-community/data-quality-insights/src/domain";
+
+export {ProductEvaluation, CriterionEvaluationResult, Product};