+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/hooks/ProductEditForm/useProduct.ts
@@ -1,8 +1,0 @@
-import {useSelector} from 'react-redux';
-import {ProductEditFormState} from "../../store";
-
-const useProduct = () => {
-  return useSelector((state: ProductEditFormState) => state.product);
-};
-
-export default useProduct;